Micro adjustments

+++ FinalDeliverable2/Deliverable 2/app/src/main/java/com/example/myapplication/DeleteClass.java
@@ -12,12 +12,14 @@ import android.widget.Toast;
 public class DeleteClass extends MainActivity {
 
     EditText classname;
+    String name;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_delete_class);
 
+        name = getIntent().getStringExtra("name");
         classname    = (EditText)findViewById(R.id.editClassName2);
     }
 
@@ -27,14 +29,23 @@ public class DeleteClass extends MainActivity {
 
         if (res.getCount() != 0) {
 
-            boolean isDeleted = myDB.deleteClass(classname.getText().toString());
+            res = myDB.getAllClassesByClass(classname.getText().toString());
+            res.moveToNext();
+            if(!res.getString(1).equals(name))
+            {
+                Toast.makeText(DeleteClass.this, "Can't delete another instructors class", Toast.LENGTH_LONG).show();
+            }
+            else
+            {
+                boolean isDeleted = myDB.deleteClass(classname.getText().toString());
 
-            if (isDeleted == true) {
-                Toast.makeText(DeleteClass.this, "Data removed", Toast.LENGTH_LONG).show();
-                finish();
+                if (isDeleted == true) {
+                    Toast.makeText(DeleteClass.this, "Data removed", Toast.LENGTH_LONG).show();
+                    finish();
 
-            } else
-                Toast.makeText(DeleteClass.this, "Data not removed", Toast.LENGTH_LONG).show();
+                } else
+                    Toast.makeText(DeleteClass.this, "Data not removed", Toast.LENGTH_LONG).show();
+            }
         }
     }
 }
